Latest Updates to Account Details
Updated the account details page so now it loads a JSON into the details, then when you click the button the details go away and is replaced by the form with any saved values appearing in the related fields. If you successfully fill out the form, it will save the new details, remove the form, and return to the details with the new values.

Also just cleaned up some files and file names.